project/requirements.txt
Flask==3.0.3
flask-bcrypt==1.0.1
flask-cors==4.0.1
Pillow==10.4.0
torch>=2.1.0
torchvision>=0.16.0